Reverted FCLSAbundanceMapper class to expecting a non-MNF image, to avoid pitfalls
FCLSAbundanceMapper was changed (before any release) so that it takes in a reference image/ non-MNF transformed image at instantiation. This facilitated validation and other tasks. It was thought that applying the MNF transform later was simple enough. However, it is highly likely that a user will induce endmembers with an MNF-transformed image that is perhaps slightly different from the MNF-transformed image that can be created from the reference image. Therefore, to avoid serious mistakes in user application, the class is now going to take in an image (MNF-transformed or PCA-transformed or otherwise) that corresponds to the exact same mixing space as the endmember spectra provided to a function, e.g., map_abundances().