Added utils.composite() for multi-process, single-band compositing; refactored AbstractAbundanceMapper
The new composite() function displaces the old composite() (now composite2()) and the old composite2() (which is now composite3()). The proliferation of composite*() functions will have to be addressed. The new composite() function is extremely fast for compositing multiple single-band rasters with NoData elements. A new partition() function was also added to utils; this displaces the AbstractAbundanceMapper's class method partition().