Updated RMSE and forward-model validation w.r.t. NoData
Previously, NoData areas were filled in with zeros. In images that have a large amount of NoData areas, this can lead to severe underestimation of the pixel-level RMSE because a number of trivially accurate (i.e., predicted 0 reflectance, actual 0 reflectance) cases. Now, this behavior is optional; by default, NoData areas are not used in calculing RMSE.